DesktopMixed: Early out instead of throwing, when no launch Change
In DesktopMixedTransitionHandler, before this CL we would throw an
exception when we try to handle a pending launch transition that doesn't
have a launch Change. Now, instead just return false when there's no
launch Change - the original launch transition might cause a transition
that isn't a launch.

Test: start Gmail in Desktop - before this CL that would cause a crash
because of a transition for GrantPermissionsActivity.
Bug: 327428659
Flag: com.android.window.flags.enable_desktop_app_launch_transitions

Change-Id: I636d9fe856096c0f305f40f79a14930ee7013736