Fix fold/unfold clear all button in Overview stops working
RecentsView is holding an old reference of mRecentsAnimationController that is already finished and never triggers the callback informing the animation is finished. This prevents the tasks to be removed and to close Overview to go back to the Home Screen. This logic happens in createAllTasksDismissAnimation after finishRecentsAnimation callback is triggered.

Fix: 377440753
Flag: EXEMPT bugfix
Test: Manual. Instructions in the bug.
Change-Id: Ib19818be9be8557e2ce9d3d440c8c062649a91bc