[flexiglass] Remove Lockscreen scene from backstack when face unlocked
...even if the alternate bouncer becomes invisible earlier than when the
device becomes unlocked.

There's a race condition between face unlock and alternate bouncer
visibility which basically causes the successful device unlock to hide
the alternate bouncer before the code in SceneContainerStartable is hit.
This leads to leaving the Lockscreen scene in the navigation back stack
even though it shouldn't be there.

This CL fixes that issue by just replacing the Lockscreen scene in the
bottom position of the navigation back stack when the device is
unlocked, regardless of whether the alternate bouncer was visible.

Fix: 375191368
Test: unit test added; older tests pass
Test: manually verified that the bug could no longer reproduce on a
UDFPS device
Flag: com.android.systemui.scene_container

Change-Id: Ie34dd753ef947c197282c9e794aebcd5102c66d8