allow configuring no focus steal for local display
The FLAG_STEAL_TOP_FOCUS_DISABLED flag prevents secondary displays
to steal focus from other displays. This flag can currently be only
set to virtual displays.

This flag replaces and extends the perDisplayFocus flag since A14,
which before was responsible to set a focus per display and prevented
the display to get focus on directed key input - this has been
removed with A14.

As an OEM developer I want to enable local physical displays to
get the FLAG_STEAL_TOP_FOCUS_DISABLED as well to be able to replace
the former perDisplayFocus functionality and to prevent it from
ever getting top focus.

Bug: 374954685
Test: DisplayServiceTests_server_display
Change-Id: I4ca9dc2af7eae08c4642b34c89deec56ed1f3d58
Signed-off-by: Carsten Rohn <carsten.rohn@bmw.de>