[PiP2] Handle open intent and PiP transition
Whenever an open intent transition comes in
with a remote transition in its request,
that transition is being handled by DefaultMixedHandler.

We need to make sure if the current task is PiP-ing,
the enter-PiP animation is being delegated to the pip handler
along the way as well.

We also need to make sure that the config-at-end activity change
is removed from the leftover handler's TransitionInfo and sent into
PipTransition.

Bug: 382715281
Flag: com.android.wm.shell.enable_pip2
Test: start a PiP activity, then adb shell am start
<non-launcher-activity>
Test: tap a notification from another app, while in PiP app

Change-Id: I5ba59fe722dc69aba9d61b17c1373e594e305be3