Move shade to the default display when the keyguard is visible
This adds a flag in the dagger graph to decide whether to move the notification stack to the default display once the keyguard becomes visible.

It's kept as a boolean constant in the dagger graph as it's not clear what the final behaviour will be there, and we want to be able to change it easily if needed.

Bug: 362719719
Bug: 381258683
Test: StatusBarTouchShadeDisplayPolicyTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I028208cf33b167c35fce9149a0f269c63cbfce77