Use correct context when creating dialogs from QS tiles
This ensures that we launch the dialog in the display QS actually are (e.g. when the shade window is in a different display)

A linter will be added to prevent dialog creation without passing the context explicitly in shade window classes.

Bug: 362719719
Bug: 383294128
Test: InternetDialogDelegateTest, UserSwitchDialogControllerTest, DataSaverTileTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Iadc5255f33f5f87b673a33352efd8340a705d365