Add MR2ServiceProvider-provided system media routes
These are routes that support routing system media, and whose properties
are similar to system-provided routes (like Bluetooth, or Built-in
Speakers).

This change also adds them to the list of transferable routes of the
system routing session, but does not implement transfers in order to
simplify the review. Transfers are coming in a following CL.

Bug: b/362507305
Test: atest CtsMediaBetterTogetherTestCases CtsMediaHostTestCasts
Flag: com.android.media.flags.enable_mirroring_in_media_router_2
Change-Id: I5950914a8387ea579d22f1bdd2c526f34b6c0758