Only apply IME insets to the floating window that has focus
In desktop windowing mode: When there were several windows of the same app with dialogs open, and one was showing the IME, the insets were applied to all of the windows/dialogs. Although, only the IME input target should receive them. This CL fixes that by checking if the current target is the IME input target and has the IME also requested to be visible.

Test: Open desktop windowing mode (two windows with dialogs open); show/hide IME on one dialog
Fix: 368444466
Flag: android.view.inputmethod.refactor_insets_controller
Change-Id: Id376f73b7c3b0315834e9be4eecc21a826db85fc