Allow dialogs to be created from the shade when on external displays
This introduces ShadeDialogContextRepository, responsible of creating a dialog context for the shade every time it moves to a separate display.

This is needed for several reasons:
- Just using the shade context would be wrong, as it is a TYPE_NOTIFICATION_SHADE window context that doesn't allow dialogs to be added
- We need to create the context before it is needed, in the background. This is expensive, and if we do it when needed on demand we might delay dialog creation on the main thread, generating jank.

This only uses the class in a couple of dialogs: more will be refactored in follow ups.

Bug: 362719719
Bug: 383294128
Test: ShadeDialogContextRepository
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I48d5e17c5530523e37c7e1a31935ac5ee1867883