[flexiglass] Fix the shade bottom paddings finally
Make the StackCutoffGuideline to be the same position as the bottom of
the scrolling stack, when scrolled all the way to the bottom, and make
some cleanups along the way.

Instead of passing in a boolean for shouldReserveSpaceForNavBar,
pass in the padding directly to the NotificationScrollingStack,
and be explicit about where notification stack paddings come from.
Use the horizontal stack paddings to constraint the
NotificationPlaceHolder, instead of adding extra space to the stack
height.

Bug: 380210512
Test: check the bottom padding of the stack in all (single, split, dual) shades
	- empty shade (no notifications)
	- few notifications (not enough to have a full height shade)
	- lots of notifications (enough to scroll the shade)
Flag: com.android.systemui.scene_container

Change-Id: I2f94c0c4203784e04fa32e0e1223c328213fb050