Move some more task-view management out of taskview impl [2/N]
Continuing to move high-level control out of taskview impl.

This focuses on combining removeTaskView with closeTaskView
since the usages were the same but obfuscated a little. In
doing so, also renamed the existing add/remove TaskView in
TaskViewTransitions to register/unregister since their only
responsibility was registering/unregistering with the
repository.

Also moved moveTaskViewToFullscreen (which is behind a flag
anyways)

This is also mechanical, but not verbatim (required a little bit
of re-arrangement, so separated it to make it easier to verify
that code was just moved.

Bug: 384976265
Test: refactor only, so existing tests
Flag: EXEMPT mechanical refactor
Change-Id: I6849311e07dc2bc2dfcd361ac81b4a3cc43d0114