Always compose both swipeToScene() in STLImpl
This CL ensures that the ModifierNodeElement associated to each swipe
orientation is composed even if the underlying DraggableHandler is not
enabled (because there is no user action in that orientation). This
allows to prevent an active SwipeToScene modifier from being
updated/interrupted during a transition.

Bug: 378470603
Test: atest SwipeToSceneTest
Flag: com.android.systemui.scene_container
Change-Id: Ia53940b39864c255a3f9a8052ead3e074ba6de0f