[sb] Expose a top-level `shouldHomeStatusBarBeVisible` flow
When inflating the top-level StatusBarRoot's AndroidView, everything is
visible at first and there is no way to hide it until all of the
view-binders start up.

The fix here is to allow for StatusBarRoot to watch for the top-level
visibility and hide the entire android view. Then when that value
changes, we can update it.

Also, add the case of HUN, which does require the status bar to be
showing.

Test: `adb shell kill com.android.systemui` with the flag on
Test: HomeStatusBarViewModelImplTest
Bug: 364360986
Flag: com.android.systemui.status_bar_root_modernization
Change-Id: I04cb436c1d2b63f4e24ab9588757be5941728c16