Dispatch pending events outside a lock
This minimizes the time that dispatchPending() holds the mCallback
lock.  mPendingEvents (which is guarded by mCallback) is copied to an
array inside the lock.  Any pending events are then transmitted
outside the lock.  This is done so that, if the client is frozen
during transmission (unlikely), the binder callback will not be
blocked.

Flag: com.android.server.am.defer_display_events_when_frozen
Bug: 298055811
Test: atest
 * DisplayServiceTests
 * CtsDisplayTestCases
Change-Id: Iaeedd8d67caba44dd0a0517316f39bb8d756ea17