[flexiglass] Initial isAutoConfirmFeatureEnabled to correct value
Up until now, we avoided initializing this to the real value to avoid
making thread-blocking calls. This CL changes that so we can get that
flow to have the proper value from the start. This is needed to fix the
attached bug.

Fix: 390028136
Test: manually verified that, with autoconform on and after a reboot,
the backspace and authentication buttons don't briefly appear on the
same first frame of the PIN bouncer just to disappear again
Flag: com.android.systemui.scene_container

Change-Id: I368590c03d7e2fd492d41e5808c1bd7627c206cc