Advance notificationsDismissPrunedSummaries to nextfood by removing the flag.
This flag has been in trunkfood full for over a month with no issues. It was incorrectly defined as a feature flag, which makes advancing to nextfood difficult.  Our options are to convert it to a new flag which we then need to roll out again, or to simply remove the flag and enable the code.  Since we have already ramped safely, it is safe to simply inline the flag.

Bug: 355967751
Flag: EXEMPT bugfix
Test: atest SystemUITests
Change-Id: I248a3f01eb3993b04ee6bb88f103671ffb6bb073