Add Transition for convert-task-to-bubble
This adds BubbleTransitions infrastructure for managing the
lifetime of a bubble-transition. Bubbles have multiple stages
to their lifecycle so this encapsulates a logical operation
into a single object.

While a transition is ongoing for a bubble, it marks itself
as a "PendingTransition" in the bubble. This is to support
coordination (since the call-graph goes through bubble
internals).

In this case (covert-to-bubble), a ConvertToBubble object is
created for every convert-to-bubble event. It coordinates with
bubble inflation, shell-transitions, and launcher-roundtrip.

Additionally, some infrastructure was set-up for associating
a bubble with a task (rather than activity). This is in the
form of a TaskInfo in Bubble and corresponding key-type.

Since convert-to-bubble is effectively "expanding a new bubble",
there is a lot of overlap with expand animation, so some of
BubbleBarLayerView was refactored to share.

A convertToBubble animation was added (using SizeChangeAnimation
ported from core).

Bug: 384976265
Test: BubbleTransitionsTest
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Change-Id: I25fde9b0a3b4b69eaad57724febab1103f8b7a0b