Add a convert-from-bubble transition
This adds convert-from-bubble transition support. Unlike the
to-bubble transition, this one does NOT actually do an animation,
it is just the necessary "setup" logic required to synchronize
taking the task out of the taskview before dispatching to a
handler which will play the actual animation. The basic principle
is that the windowing-feature that "accepts" the task is the one
that animates it.

Because bubbles use taskview and taskviews are view-hierarchys,
removing the task from the taskview seamlessly is non-trivial.
In this case, there's a `pluck` utility added which posts
reparenting on the taskview's choreographer and then hides the
taskview. Once that happens, it can pass the task-surface onto
another transition handler.

This also introduces a bubbles-specific TaskViewController so
that it can implement moveTaskViewToFullscreen with this
bubble-specific transition.

Bug: 384976265
Test: BubbleTransitionsTest
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Change-Id: I5edd380200a9839c75e37a87fb9a244b62df702b