Scrim alpha changes for bouncer
Updates the default scrim alpha to the bouncer background alpha. This makes sure that the current scrim transition alpha animations work as expected without overshooting the bouncer alpha while transitioning to the bouncer.
Setting the alpha in BOUNCER or BOUNCER_SCRIMMED state is not sufficient because while we are transitioning from KEYGUARD/SHADE_LOCKED -> BOUNCER we always interpolate the alpha up to defaultScrimAlpha and then set it to the alpha from BOUNCER state.

CUJs affected
1. Front scrim
  - Any scrimmed bouncer invocation, i.e tapping on any UI element that requires authentication and ends up showing the bouncer.
  - Tapping on locked out SFPS
  - Both large and small screen devices
2. Notif scrim
  - On small screen devices, notif scrim fades in behind the bouncer when we swipe up to go to bouncer
  - On large screen devices, notif scrim does not fade in behind the bouncer.
3. Behind scrim
  - On large screen devices, behind scrim fades in behind the bouncer when we swipe up to go to the bouncer.

Bug: 381263600
Test: verified the above CUJs with the flag on
Flag: com.android.systemui.bouncer_ui_revamp
Change-Id: Ic774b387ba9db610a34c862d3269170635243c16