Add catch block for ClassCastException in areBundlesEqual.
We've now ween a couple cases where a ClassCastException is happening in
the areBundlesEqual method.  It is not clear what is causing that issue
so instead of just silently catching this exception, we'll log which key
is causing the problem so we can investigate further.

Test: Compile; run tests -- this is not easily repro.
Flag: NONE crash logs only.
Fixes: 390386486
Bug: 372319846
Change-Id: I96c5d08617487add9ba83de80144ae8b2f74782d