Hide IME only after bubble drag started
We currently hide the IME in the onDown of the touch gesture. If we're
just tapping to switch bubbles, hiding the IME interferes with the
switch animation. This change hides the IME when a move gesture starts.

Demo: http://recall/-/bJtug1HhvXkkeA4MQvIaiP/dJXt0FwoLKV8aEuMGd5dJe

Flag: EXEMPT bugfix
Fix: 384838831
Test: manual
      - Have 2 floating bubbles
      - Expand and show IME
      - Switch to the other bubble
Change-Id: Iea5a7087e9604b18386c6c292d9af84a90f0301e