Do not freeze system_server
In early testing, freezing the current process generated an error but
did no harm.  Now it does harm.  This CL disables freezing because of
an ANR timeout if the target pid is 0 (often passed in to signify
"self") or the pid of system_server.

Flag: com.android.server.utils.anr_timer_freezer
Bug: 391491604
Test: atest
 * FrameworksServicesTests:AnrTimerTest
 * FrameworksServicesTests:com.android.server.am
 * FrameworksMockingServicesTests:com.android.server.am
Change-Id: I78189987f4dee398a02f6ca50df9d4e1a83b5419