Extract animators into separate methods
The plan to fix the NPE is to not store the animators as fields, store only the combined animator set, cancel and re-instantiate and start it every time.
This CL extracts all usages of the cameraProtectionAnimator field to be a parameter for all methods that use it.

Bug: 392032258
Test: NA
Flag: EXEMPT extracting code into separate methods.
Change-Id: Ib0ce136d0d6c1dff8623cf6ddb217f4b249ccdbd