Adding in first version of Advanced Protection Service USB Data Protection Feature
Adds a hook to the Advanced Protection Service, controlling the USB Data protection feature: when the feature is enabled, when the phone is locked, any plugged in USB devices that do not have an active data connection are forcibly disabled.

Flag will remain in teamfood for testing until a first release of APM is in stable state. And adjustment to existing CTS tests is completed.

Change-Id: I58069b0db7f562e3504d6958d5a65288029fe9ef
Testing: Manually tested, follow up test would need to be done in CTS-I when infrastructure is ready.
Flag: android.security.aapm_feature_usb_data_protection
Bug: 389958463