[1/3] Rename OngoingActivityChipModel types to Active/Inactive.
See go/chip-transitions-proposal for details.

Note that I simplified it a bit compared to that design, by keeping
the binary structure and just adding a `hidden` field to the `Active`
class. This makes the migration easier as we're not adding the extra
layer shown in the doc.

Bug: 273205603
Flag: EXEMPT rename only
Test: atest CallChipViewModelTest CastToOtherDeviceChipViewModelTest NotifChipsViewModelTest ScreenRecordChipViewModelTest ShareToAppChipViewModelTest ChipTransitionHelperTest OngoingActivityChipsViewModelTest OngoingActivityChipsWithNotifsViewModelTest HomeStatusBarViewModelImplTest

Change-Id: I6cad9d4210fbc22a96398dfd62dcf505a2eeea20