[SB] Only show composable chips when there is enough space
This CL adds custom layout modifiers for OngoingActivityChip
and ChipContent to ensure that they are only shown if there is enough
width available in the status bar.
For timers, the entire text needs to have enough space.
For text, at least 50% of it must be visible.
Also, consolidates existing ChipText logic into ChipContent so we
do not need to make any of the modifiers public.

Bug: b/390530157
Flag: com.android.systemui.status_bar_chips_modernization

Test: Make sure the chips are sized correctly based on available width.
Test video provided in b/390530157 comment#4

Change-Id: I74ca45f96a3722482c7ae5282cccb7e895423f6e