Check if getParameters reply contains the required key
For getAdditionalOutputDeviceDelay and getMaxAdditionalOutputDeviceDelay
the method expects that the AudioSystem.getParameters() returns a reply
in a form: `<key>=<value>`. If the optional parameter is unsupported
it expects that reply string is constructed as `<key>=""` but instead,
it might receive empty string that gives
StringIndexOutOfBoundsException.

Bug: 393338993
Test: run cts -m CtsMediaTestCases -t android.media.cts.AudioManagerTest
Change-Id: I6bc6043f60092e278cc00e3c7c5b5f4974e8e9a5
(cherry picked from commit 2c303494532d007f9dd1494c82a6a8fd53118bee)