add jarjar_prefix for framework-location
framework-location is a BootClassPath jar on the platform. It doesn't
depend on framework.jar, however it uses the flags from the
framework.jar. Flags in framwork.jar were renamed, and some of them will
be renamed. So framework-location needs to rename these flags in the
same way to use the flags in the framework.jar

Test: presubmit
Bug: 390062054
Change-Id: I20b48cdcf2d52b19fc167a9f893c2f777cee322f